1. RegisterMemberDTO 분리
